+++ src/components/atoms/UserList/UserList.jsx
@@ -1,42 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import styled from 'styled-components';
-import Button from '../Button/Button';
-
-const UserListItem = styled.li`
-    height: 50px;
-    margin: 16px;
-    display: flex;
-    background-color: #fff;
-`;
-
-const UserImg = styled.img`
-    margin-right: 12px;
-    border-radius: 70%;
-`;
-
-const UserNickname = styled.strong`
-    margin: 6px 12px;
-`;
-
-const UserIntroduction = styled.p`
-    margin: 6px 12px;
-`;
-
-const FollowBtn = styled(Button)`
-    min-width: 56px;
-    margin: 0;
-    && {
-        margin-left: auto;
-        align-self: center;
-    }
-`;
-
-const UserDsc = styled.div`
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-`;
+import UserListItem from '../UserListItem/UserListItem';
 
 export default function UserList({ imgSrc, children, intro }) {
     // 팔로워 리스트 가져오기
@@ -65,19 +29,6 @@ export default function UserList({ imgSrc, children, intro }) {
     }, []);
 
     return (
-        <>
-            {followers.length
-                ? followers.map((follower, i) => (
-                    <UserListItem key={i}>
-                        <UserImg src={API_HOST + follower.image} alt="팔로워 리스트 프로필" />
-                        <UserDsc>
-                            <UserNickname>{follower.accountname}</UserNickname>
-                            <UserIntroduction>{follower.intro}</UserIntroduction>
-                        </UserDsc>
-                        <FollowBtn size="small" children="팔로우" />
-                    </UserListItem>
-                ))
-                : null}
-        </>
+        <>{followers.length ? followers.map((follower, i) => <UserListItem follower={follower} key={i} />) : null}</>
     );
 }
